Added function name preview feature
