+++ public/notes/hw2-and.md
@@ -1,0 +1,3 @@
+and2(a,b)
+and3(a,b,c)
+andN(array)
