+++ public/notes/hw2-fract.md
@@ -1,0 +1,1 @@
+fractionString(n,d)
