+++ public/notes/hw2-song.md
@@ -1,0 +1,1 @@
+generateSong()
