+++ public/notes/hw2-triangle.md
@@ -1,0 +1,1 @@
+letterTriangle(n)
