+++ views/index.jade
@@ -3,13 +3,19 @@ extends layout
 block content
   h1= title
   form(method='POST')
-    select(name='test')
+    fieldset.exports
+      legend Functions to be Tested
+      - var attributes = { 'autocomplete': 'off', 'cols': '28', 'rows': '4', 'spellcheck': 'false' }
+      textarea#exports(name='exports' readonly)&attributes(attributes)
+    select#testSelector(name='test')
       //- TODO: make the optgroups and options autogenerated
       optgroup(label='Week 2')
-        option(value='hw2') Cards
-        option(value='hw2-fract') Fractions
-        option(value='hw2-song') Song lyrics
-        option(value='hw2-and') && simulator
+        option(value='hw2-fract') 1: Fractions
+        option(value='hw2-triangle') 3a: Letter triangle
+        option(value='hw2-song') 3b: Song lyrics
+        option(value='hw2-and') 4: && simulator
+        option(value='hw2-cards') 5a,b: Cards
+    //- Hide until future weeks:
       optgroup(label='Week 3')
         option(value='hw3-arrays') Arrays      
         option(value='hw3-cards') Cards
@@ -22,10 +28,10 @@ block content
         option(value='hw5-cards') Cards
         option(value='hw5-deque') Deque
         option(value='hw5-secrets') Secrets
+    button#submit(formaction='/done') Test!
+    br
     fieldset.code
       legend Enter your code here
       - var attributes = { 'autocomplete': 'off', 'cols': '80', 'rows': '30', 'spellcheck': 'false' }
       textarea#code(name='code')&attributes(attributes)
-    br  
-    button#submit(formaction='/done') Test!
     div#mocha
