+++ views/layout.jade
@@ -10,6 +10,7 @@ html
       ga('create', 'UA-65533983-2', 'auto');
       ga('send', 'pageview');
     script(src='https://ajax.googleapis.com/ajax/libs/jquery/2.1.3/jquery.min.js')
+    script(src='/js/preface.js')
     link(rel='stylesheet', href='/stylesheets/style.css')
     link(rel='stylesheet', href='/mocha/mocha.css')
   body#main
